Add reviewer recommendations

+++ db/migrate/20220113175240_add_comments_counter_to_post.rb
@@ -1,5 +1,5 @@
 class AddCommentsCounterToPost < ActiveRecord::Migration[7.0]
   def change
-    add_column :posts, :comments_counter, :decimal
+    add_column :posts, :comments_counter, :integer
   end
 end
