+++ db/migrate/20220113175312_add_likes_counter_to_post.rb
@@ -1,5 +1,5 @@
 class AddLikesCounterToPost < ActiveRecord::Migration[7.0]
   def change
-    add_column :posts, :likes_counter, :decimal
+    add_column :posts, :likes_counter, :integer
   end
 end
